+++ js/main.js
@@ -1,3 +1,2 @@
 import './big-picture.js';
 import './picture.js';
-import './comments.js';
